services/auth-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
PyJWT==2.8.0
passlib[bcrypt]==1.7.4